Implemented moveHorizontal and changed push parameters
moveHorizontal moves a token on the row indicated by the die roll forward one place. If there is more than one stack to choose from to move forward it gives the choice to the user on what stack they would like to move.
The push function in the stack library no longer takes a player as a parameter but rather the colour of a token. This simplifies code in moveHorizontal.
Also removed moveToken as it was unnessesary.